Add Terraform configuration options
These configuration options allow this function to be built through terraform and is designed to be used as a `module` in a larger terraform deployment. It creates a deployment package for the lambda function and updates it if it's hash does not match the previously deployed version.

This currently does not implement the necessary layer sub-modules but functions as a proof-of-concept for this method of managing the SFR data pipeline